requirements.txt
aiodns==3.0.0
aiohttp[speedups]==3.8.1
aiosignal==1.2.0
ansi2html==1.7.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
blake3==0.3.1
brotli==1.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cython==0.29.28
defusedxml==0.7.1
deprecated==1.2.13
dunamai==1.11.0
ecs-logging==1.1.0
elastic-apm==6.8.1
elastic-enterprise-search==7.17.0
elastic-transport==7.16.0
elasticsearch[async]==7.17.1
frozenlist==1.3.0
funcparserlib==1.0.0a0
get-version==3.5.4
hiredis==2.0.0
hy==1.0a4
hyrule==0.1
idna==3.3
jxlpy==0.9.1
lxml==4.8.0
multidict==6.0.2
namedthreads==1.0
numpy==1.22.3
orjson==3.6.7
packaging==21.3
pillow==9.0.1
pycares==4.1.2
pycparser==2.21
pycryptodome==3.14.1
pycurl==7.45.1
pyjwt==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-levenshtein==0.12.2
redis[hiredis]==4.2.0
rply==0.7.8
setuptools==61.2.0
six==1.16.0
soupsieve==2.3.1
tornado==6.1
typing-extensions==4.1.1
urllib3==1.26.9
uvloop==0.16.0
wrapt==1.14.0
yarl==1.7.2
